Permit file:// URLs for disag PD
The namespaced IPC sockets used currently do not work well in OCI
containers (#63) due to namespacing approach but filesystem-based
sockets are fine.

Implement `--pd-url` parsing to determine if the scheme is file^tcp
and use the path attribute of the file-based url to deploy the PD
sockets for disag processing (leaving prior code paths the same).

For Docker/Podman, this allows use of a shared volume/directory
between the containers to connect to each-other's sockets. Network
enabled compute environments sharing fast storage can utilize this
facility to avoid the copy-to-CPU effect of using the TCP socket
over POSIX-compatible FS exports.